# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Menus",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/restaurants": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Mostrar restaurantes",
                "description": "Listado de restaurantes",
                "operationId": "edd56765710407a7ca2f654206d81a79",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los restaurantes."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Crear restaurante",
                "description": "Crear un restaurante",
                "operationId": "1e8bf99418f620d1c2979ae1302825be",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crea un restaurante."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/restaurants/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Mostrar restaurante",
                "description": "Detalle de un restaurante",
                "operationId": "5d65be63943bc4d559efb05d614b8f42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del restaurant",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar detalle de un restaurante."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Crear restaurante",
                "description": "Crear un restaurante",
                "operationId": "07717cf6d215bf69168fe046e5161481",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del restaurant",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crea un restaurante."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Eliminar restaurante",
                "description": "Eliminar un restaurante",
                "operationId": "be788fad15027bec70347318f364ef82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del restaurant",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un restaurante."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}